Add a working devcontainer configuration.
This config will automatically pick up and install the devShell
dependencies from `flake.nix` when the container is created or started.
If you are editing the flake.nix in the meantime, you'll need to open a
terminal and run `update-content.sh`. You may need to reload the remote
vscode window in order for this to take effect, but I tested that it
works with the golang extension.