Removes pre-2.7 flake compatibility (per #10)
As reported in https://github.com/peterldowns/nix-search-cli/issues/10
the cargo-culted warning about flake compatibility is showing up for
users who are not on pre-2.7 nix. Additionally, running `nix flake
check` shows warnings about the existence of the `defaultApp`,
`defaultPackage`, and `devShell` keys that are only defined for
compatibility purposes.

At this point, feels safe enough to remove the warning. If someone
complains because they are on a very old version of flakes I can easily
revisit this decision.

For background information on the flake schema change as of 2.7, see:

- https://nixos.org/manual/nix/stable/release-notes/rl-2.7.html#release-27-2022-03-07
- https://github.com/NixOS/nix/issues/5532#issue-1049704766