Update flake.lock to fix issue on MacOS Tahoe (#25)
Running this as a flake on tahoe produces the error: 
```
missing LC_UUID load command in /nix/store/z7syn0ajm5x21cri8vvbqqdbs7wqz97l-nix-search-v0.3-/bin/nix-search
```
Which is due to a [change in
behavior](https://developer.apple.com/documentation/technotes/tn3178-checking-for-and-resolving-build-uuid-problems)
on tahoe and is referenced in [this stackoverflow
post](https://developer.apple.com/documentation/technotes/tn3178-checking-for-and-resolving-build-uuid-problems)
as something that is fixed in newer versions of go.

Updating the reference to nixpkgs seems to fix the issue.

Can be reproduced (on tahoe) with `nix run` on the previous lockfile,
and I have verified that it is fixed with the new lockfile